tried to merge with master from 01.04.2020 resolved some but not all isses

+++ src/main/java/kernbeisser/CustomComponents/ShoppingTable/ShoppingCartController.java
@@ -10,7 +10,7 @@ public class ShoppingCartController implements Controller {
     private ShoppingCartModel model;
 
 
-    public ShoppingCartController(int userValue, int userSurcharge) {
+    public ShoppingCartController(double userValue, int userSurcharge) {
         model = new ShoppingCartModel(userValue, userSurcharge);
         view = new ShoppingCartView(this);
         refresh();
